+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Change Log
 
-## [1.0-alpha04] - 2018-04-08
+## [1.0-alpha04] - 2018-04-14
 - android support library 27.1.1
 - several API changes:
+    - renamed `FabWithLabelView.setOptionFabSelectedListener()` to `FabWithLabelView.setOnActionSelectedListener()`
+    - renamed `FabWithLabelView.isLabelEnable()` to `FabWithLabelView.isLabelEnabled()`
     - added `SpeedDialActionItem.Builder(@IdRes int id, @Nullable Drawable d)`
     - added `SpeedDialView.setOnChangeListener(OnChangeListener l)`
     - removed `SpeedDialView.setMainFabOnClickListener(OnClickListener l)`
-    - renamed `FabWithLabelView.setOptionFabSelectedListener()` to `FabWithLabelView.setOnActionSelectedListener()`
     - renamed `OnOptionFabSelectedListener` to `OnActionSelectedListener`
     - renamed `SpeedDialView.setOptionFabSelectedListener()` to `SpeedDialView.setOnActionSelectedListener()`
     - renamed `SpeedDialView.addAllFabOptionItem()` to `SpeedDialView.addAllActionItems()`